refactor: rename headless_personality* to personality and personality_routes
