+++ src/reachy_mini_conversation_app/console.py
@@ -40,9 +40,9 @@ from reachy_mini_conversation_app.config import (
     refresh_runtime_config_from_env,
 )
 from reachy_mini_conversation_app.startup_settings import read_startup_settings, write_startup_settings
+from reachy_mini_conversation_app.personality_routes import mount_personality_routes
 from reachy_mini_conversation_app.audio.startup_config import apply_audio_startup_config
 from reachy_mini_conversation_app.conversation_handler import ConversationHandler
-from reachy_mini_conversation_app.headless_personality_ui import mount_personality_routes
 
 
 try:
